mlservice/requirements.txt
flask
mne
numpy
catboost
scipy
python-dotenv